Refactor StaticAbilityCantPhaseIn and StaticAbilityCantPhaseOut into StaticAbilityCantPhase
