+++ forge-game/src/main/java/forge/game/card/Card.java
@@ -5468,11 +5468,11 @@ public class Card extends GameEntity implements Comparable<Card>, IHasSVars {
     }
 
     private boolean switchPhaseState(final boolean fromUntapStep) {
-        if (isPhasedOut() && StaticAbilityCantPhaseIn.cantPhaseIn(this)) {
+        if (isPhasedOut() && StaticAbilityCantPhase.cantPhaseIn(this)) {
             return false;
         }
 
-        if (!isPhasedOut() && StaticAbilityCantPhaseOut.cantPhaseOut(this)) {
+        if (!isPhasedOut() && StaticAbilityCantPhase.cantPhaseOut(this)) {
             return false;
         }
