--- forge-game/src/main/java/forge/game/staticability/StaticAbilityCantPhaseIn.java
@@ -1,32 +1,0 @@
-package forge.game.staticability;
-
-import forge.game.Game;
-import forge.game.card.Card;
-import forge.game.zone.ZoneType;
-
-public class StaticAbilityCantPhaseIn {
-
-    static String MODE = "CantPhaseIn";
-
-    static public boolean cantPhaseIn(Card card) {
-        final Game game = card.getGame();
-        for (final Card ca : game.getCardsIn(ZoneType.STATIC_ABILITIES_SOURCE_ZONES)) {
-            for (final StaticAbility stAb : ca.getStaticAbilities()) {
-                if (!stAb.checkConditions(MODE)) {
-                    continue;
-                }
-                if (applyCantPhaseIn(stAb, card)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    static public boolean applyCantPhaseIn(StaticAbility stAb, Card card) {
-        if (!stAb.matchesValidParam("ValidCard", card)) {
-            return false;
-        }
-        return true;
-    }
-}
