--- forge-game/src/main/java/forge/game/staticability/StaticAbilityCantPhaseOut.java
@@ -1,32 +1,0 @@
-package forge.game.staticability;
-
-import forge.game.Game;
-import forge.game.card.Card;
-import forge.game.zone.ZoneType;
-
-public class StaticAbilityCantPhaseOut {
-
-    static String MODE = "CantPhaseOut";
-
-    static public boolean cantPhaseOut(Card card) {
-        final Game game = card.getGame();
-        for (final Card ca : game.getCardsIn(ZoneType.STATIC_ABILITIES_SOURCE_ZONES)) {
-            for (final StaticAbility stAb : ca.getStaticAbilities()) {
-                if (!stAb.checkConditions(MODE)) {
-                    continue;
-                }
-                if (applyCantPhaseOut(stAb, card)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    static public boolean applyCantPhaseOut(StaticAbility stAb, Card card) {
-        if (!stAb.matchesValidParam("ValidCard", card)) {
-            return false;
-        }
-        return true;
-    }
-}
